pyproject.toml
[project]
name="pymongo-database-dashboard"
version="1"
dependencies=[
  "jupyter-dash",
  "numpy",
  "pandas",
  "matplotlib",
  "dash-leaflet",
  "pymongo"
]